Switch back to simple and linear probing
This gives better cache locality of secondary and latter probes.

On the other hand, the initial probe has no cache locality and could be
improved to have more entropy; this is something for the future,
however.

Btw, this beat haxe.ds.IntMap's khash for time (mono + Linux) spent on a
simple insertion test.  Storage and reliability, however, are still far
better on that implementation.